Eliminate evaluate() and coalesce evaluation blocks
Allocating a lambda is not free, as well as entering and exiting the
evaluation block. Of course, with blocks there's a tradeoff, because it
is easier to accidentally call a guarded method of an instance from
another one, causing a false positive circular evaluation failure.

However, microbenchmark shows a significant improvement, and some
performance tests look noticeably better with larger blocks.